oci: extract IsOCIPath out to a file that runs on all platforms
