+++ internal/oci/oci_wasi_tinygo.go
@@ -8,11 +8,6 @@ import (
 	"errors"
 )
 
-// IsOCIPath checks if a given path is an OCI path
-func IsOCIPath(path string) bool {
-	return false
-}
-
 func PullWIT(ctx context.Context, path string) (*bytes.Buffer, error) {
 	return nil, errors.New("OCI not supported on WASI or TinyGo")
 }
